Angular 9 compatibility: add explicit types to prevent automatic deep imports
If you do not provide an explicit return type to a function then TypeScript
will try to guess it and create a dynamic import type for the `.d.ts` files.

Often the imports incorrectly point to a path deep within a package,
rather than accessing the type via the correct public facing import.

From Angular v9, using ivy, such deep imports cause compilation
errors and will prevent this library from being used for ivy enabled
applications.

This commit fixes this problem for a component and a service that
the adf-core library exports.