AAE-35668 Sending empty string instead of null after clearing value from numeric field (#10925)
* Handles null values in number widget

Ensures the number widget correctly handles null, undefined, and empty string values, setting the field value to null in these cases.

This change prevents unexpected behavior when the user clears the input field, ensuring data consistency.

* Adds test for readonly number widget

Adds a test case to verify the behavior of the number widget when it's in readonly mode.

The test checks if the displayValue is correctly set using the decimalNumberPipe when the readOnly property is true.

* Refactors number widget tests

Updates number widget tests to use `overrideComponent` for providing mocked dependencies. This approach ensures proper isolation and avoids potential issues with shared state between tests.
Additionally, it adds a test case to verify the `displayValue` is correctly set using the mocked `DecimalNumberPipe`.

---------

Co-authored-by: Eugenio Romano <eromano@users.noreply.github.com>