AAE-35728 Add aria label to viewer loading bars (#10927)
* Adds loading state accessibility

Improves accessibility by adding screen reader support to loading indicators.

This ensures that users with disabilities are informed when the document is loading, improving the overall user experience.

* Trigger Build

* Fixes accessibility issue on loading spinner

Removes the redundant `attr.` prefix from the `aria-labelledby` attribute in the loading spinner.
This resolves an accessibility issue where the spinner's label was not correctly associated, enhancing the user experience for assistive technologies.

* Trigger Build