AAE-36869 Handles error messages with wrong JSON format (#11078)
* Handles error messages with wrong JSON format

Handles cases where the error message from the backend is not a valid JSON.

If the message is not a valid JSON, it displays the error as a string.
This prevents the application from crashing when receiving error messages with wrong JSON format.

Fixes AAE-36869

* SonarCloud solution

* Displays the error message from the response

The error message is now extracted from the response body,
providing more context when process start fails.

This change ensures that the user sees the specific error
message returned by the service when a process instance
cannot be started, improving the user experience.

* Refactors start process template to use `@if` blocks

Migrates the start process component template from `*ngIf` directives to the new Angular `@if` syntax for improved readability and performance.
This change enhances the structure and efficiency of conditional rendering within the template.

* prettier