AAE-36822 Removes showNextTaskCheckbox input for the feature flag (#11066)
* Removes showNextTaskCheckbox input

Removes the `showNextTaskCheckbox` input from the `UserTaskCloudComponent` and related files.

The `showNextTaskCheckbox` input is no longer needed, instead the visibility of the 'Open next task' checkbox will depend on the canCompleteTask() method.
This change simplifies the component's API and ensures that the checkbox is always displayed when it should be.

* Adds option to open next task after completion

Introduces a new checkbox on the task screen that, when checked, automatically navigates the user to the next available task upon completion of the current one.
It enhances user experience by streamlining task completion workflows, eliminating the need for manual navigation.

* revert changes

* review