AAE-38063 Default buttons are temporarily visible before the custom buttons in workspace (#11198)
* Bonus Content: Remove warnings from console

* Fixes: Prevents display issues during form loading

Ensures the process definition card and start button are not rendered until the form is fully loaded or an error occurs during the loading process.
This prevents flickering and ensures a smoother user experience.

* BONUS content: calm down sonar cube

* Removes unnecessary error handling

Removes the error handler from the `subscribe` block of the `startProcess` method in `StartProcessCloudComponent`.
The error handling was redundant, as errors are already handled by other mechanisms.
This simplifies the code and avoids potential duplicate error handling.

* Uses `test` method for regex checks

Replaces the `exec` method with the `test` method for regular
expression checks to determine if a pattern exists within a string.

This resolves potential issues where the return value of `exec`
(an array or null) was not being properly evaluated as a boolean
condition, leading to incorrect logic execution.

---------

Co-authored-by: Fabian Kindgen <39992669+fkindgen@users.noreply.github.com>